Move methods to base class
